snowflake: parallelize building of parquet rows
There is a lot of latency to build a parquet file, we can speed things
up by doing the most expensive part of `constructRowGroup` in parallel.

The actual construction of the parquet is still expensive, especially
in the compression. We can't parallelize that with this library, but
that is the next step if this is still slow. Hopefully this will help
max out CPU in the benchmarks I'm running.

+++ internal/impl/snowflake/streaming/stats.go
@@ -68,7 +68,7 @@ func computeColumnEpInfo(transformers []*dataTransformer, stats []*statsBuffer) 
 			maxStrVal = &s
 		}
 		info[transformer.column.Name] = fileColumnProperties{
-			ColumnOrdinal:  int32(transformer.column.Ordinal),
+			ColumnOrdinal:  transformer.column.Ordinal,
 			NullCount:      stat.nullCount,
 			MinStrValue:    minStrVal,
 			MaxStrValue:    maxStrVal,
